bennoune/requirements.txt
coverage==4.4.1
Django==2.1.1
django-nose==1.4.5
djangorestframework==3.8.2
nose==1.3.7